% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob.R
\name{condprob}
\alias{condprob}
\title{Function condprob}
\usage{
condprob(xX, xY, xImpair, ProbComp = c("gt", "lt"), Exceed = c("gte",
  "lte"), ci = FALSE, alpha = 0.05, R = 100, xW = rep(1,
  length(xX)))
}
\arguments{
\item{xX}{numeric vector for conditional (independent) variable}

\item{xY}{numeric vector for dependent variable, conditional on \code{xX}}

\item{xImpair}{numeric threshold for impairment of \code{xY}}

\item{ProbComp}{chr string}

\item{Exceed}{chr string}

\item{ci}{logical to return confidence interval for estimated likelihood of impairment}

\item{alpha}{numeric for alpha level defining confidence}

\item{R}{numeric for number of bootstrap iterations to estimate confidence}

\item{xW}{numeric vector with length equal to \code{xX}}
}
\description{
This function calculates a conditional probability analysis
}
\examples{
data(jeqdata_wq)
ept<-jeqdata_wq$EPT.RICH
pct_fn<-jeqdata_wq$PCT.FN
cpa1<-condprob(pct_fn,ept,9,'lt','gte',T,R=1000)

}
