% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob_faster.R
\name{condprob_faster}
\alias{condprob_faster}
\title{Function condprob}
\usage{
condprob_faster(xX, xY, xImpair, ProbComp = c("gt", "lt"), Exceed = c("gte",
  "lte"), ci = FALSE, alpha = 0.05, R = 100, xW = rep(1, length(xX)))
}
\arguments{
\item{xX}{description}

\item{xY}{description}

\item{xImpair}{description}

\item{ProbComp}{description}

\item{Exceed}{description}

\item{ci}{description}

\item{alpha}{description}

\item{R}{description}

\item{xW}{description}
}
\description{
This function calculates a conditional probability analysis
}
\examples{
data(jeqdata_wq)
ept<-jeqdata_wq$EPT.RICH
pct_fn<-jeqdata_wq$PCT.FN
cpa1<-condprob_faster(pct_fn,ept,9,'lt','gte',T,R=1000)

}
