\name{condprob}
\alias{condprob}
\title{Function condprob}
\usage{
condprob(xX, xY, xImpair, ProbComp = c("gt", "lt"), Exceed = c("gte",
  "lte"), ci = FALSE, alpha = 0.05, R = 100, xW = rep(1, length(xX)),
  minSamp = length(xX))
}
\arguments{
  \item{xX}{}

  \item{xY}{}

  \item{xImpair}{}

  \item{ProbComp}{}

  \item{Exceed}{}

  \item{ci}{}

  \item{alpha}{}

  \item{R}{}

  \item{xW}{}
}
\description{
This function calculates a conditional probability analysis
}

